Update folder equivalence: see full description
Equivalence between folders (between past and present versions of the same folder, between different folders) now checks for identical folderName on top of equivalence between cards. Both conditions need to be true for a folder to equal another.

In the process of making this change, constructors of CardFolder and ModelManager have been modified to require a folderName when creating a new folder from nothing.

The default behaviour of MainApp to instantiate a sample cardfolder has now been removed; if none is read from storage, no cardfolder will be created on instantiation.

Tests have also been heavily modified to reflect the above changes.