v1.1Sort added the global change, sort
Implemented the sort command that the GUI will register the
sort command and output the success message. The changes made
will have an global impact on other commands. For instance, the
list command will have a different output due to the sort. In current
version, the sorting will not perform any action. However, in future
versions, we will then implement the sort.

Resolves: #36
See also: #67