(V1.2) Editing JsonAdaptedPerson & PersonBuilder
This commit has a couple of changes. The main focus of the changes
surround JsonAdaptedPerson, that will soon be refactored to
JsonAdapetedPdf. The usage of the class extends deep into the test
files territory and thus is requiring major work overs. This commit
WILL NOT PASS Travis tests. Further rework and update from Model
side is necessary for smoother proceedings.

This commit is simply to push Storage edit progress and keep the
team with updated code.