(V1.2) UserGuide Formatting Change
The UserGuide Table of Contents has been put to the left and will move
with the user as they scroll. This is to allow for ease of access to
various content with a few button clicks.

Also, there was an issue with the FXML browser such that it did not
render the guide correctly. Hence the loading of the HTML Page has been
offloaded to the OS. Whichever browser is default in the OS will load
the UserGuide when the user clicks Help/F1.

Resolves: #109