Resolve equivalence conditions for folders
Folder equivalence is determined by folder name. However, when comparing VersionedCardFolders, an additional condition is enforced, where past states need to have the same cards as well on top of having the same name. This is so that VersionedCardFolders will be different if they have different sets of cards in their past states.