Some structural changes to PrimitiveRequirement and Condition class. Now condition stores a Pattern, pattern, and an int, minToSatisfy (signifiying the minimum number of unique modules must match the pattern in order to satisfy the condition). Removed conditionName attribute from Condition class. PrimitiveRequirement stores a list of Conditions.
Also changed classes and tests that depended on the new PrimitiveRequirement and Condition classes. Created new Json files for course due to the structural changes above.

Realised that completing at least 40 modules is not really a Graduation requirement since NUS counts modular credits instead of module count. But we can leave it for now.