(v1.4) Included WARNING in Userguide due to misuse
Some of the reported bugs were due to users carrying out manual actions
to the files after adding them to the app. This however is no the
intended use of the application and hence a warning has been posted on
the user guide that misuse or unforeseen manual actions may result in
the application misbehaving.

Resolves: #237 & #230