(v1.4) done and remove throw error if misused.
Previously, if a deadline did not exist in a pdf, the use of done or
remove would not throw an error and simply do nothing. Hence the bug has
been resolved by getting the DeadlineCommand to throw an error if the
user wishes to use the done or remove prefix with a pdf that does not
already contain a deadline.

Resolves: #232