(v1.4) SizeTest In Progress
From prior testing, the NameTest already works (implementation in v1.3
of pdf++). SizeTest however was incomplete. Fixed and revamped the test
format to get it to work. SizeTest now works completely HOWEVER. From
the trial demo, it was highlighted that we should make the size value
more "readable" and this will require several changes in the Size,
JsonAdaptedPdf and other class files. Proceeding to work on those
changes now.

Track Changes: #217