(v1.4) SizeTest Complete
SizeTest is now comprehensive and complete. This commit contains some
checkstyle fixes for Travis Erros that appeared in the previous commit
(Pdf.java javadoc missing). Also implemented the use of static variables
for threshold and prefix values so that any changes made will be easily
ported over to the test cases. One minor change was to make the regex
check in the Size.java a private static variable as per IntelliJ's
recommendation.

Track issue: #217