Update UG
Remove redundant images as the user already has the GUI running and
there is really no point in reproducing an image just for the sake of
it, besides that, the user should be able to follow the instructions
given in the quick start relatively easily. It should have been written
in such a manner as to encourage them to read on and play the game,
instead of giving them "training wheels" to play this already-simple
game.